requirements.txt
asgiref==3.6.0
Django==4.2.1
Pillow==9.5.0
python-decouple==3.8
sqlparse==0.4.4
